Documentation on prefixes, improved performance.
When looking for included files, a lot of file names were queried using
the exist() function of Path. For large projects, this could take a
considerable amount of time (at least on Windows, where I tried). Now,
mymake asks for all files in a directory and puts them in an efficient
cache internal cache. This cache is structured so that answering the
questions mymake is after, is a lot quicker than asking the OS for all
possibilities.