Improved diagnostic messages.
The --debug flag now takes either a number or the name of the messages
to show. Ie. one can now use mm --debug verbose to enable messages. All
debug messages above normal will also tell which level they originate
from, so that it is easier to discover the different levels.

Added the new level PEDANTIC to allow for a level that spots suspicious
things which are not always errors that one can use to easily spot
common mistakes. Currently, I added a warning for including files
outside the working directory. These were previously silently
ignored. This is currently the only warning for PEDANTIC, but more will
come as they are needed.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>