Synchronize during process shutdown.
Previously a process with much output could get its output truncated due
to mymake closing the pipe early. This was caused by the termination
signal arriving before mymake could flush the buffer, and could be
easily seen on Windows. Now, we make sure to drain the pipe before
closing it on our side. This makes our outputmgr slightly less general,
but it is more than adequate for what mymake needs.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>