Fixed a threading issue causing deadlock.
Since Mymake was sometimes too eager to close pipes, the OS could re-use
the handle values of those pipes before Mymake was actually done with
using the handle as an unique identifier (no accesses were made, but we
still assumed the handle was unique). This is now fixed by delaying
closing of pipes until we do not need their handles as identifiers
either.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>