Reworked the output manager a bit to avoid strange lock-up.
Now, we don't ensure that OutputMgr::removePipe flushes all output on the
spot. For some reason, some pipes seem to linger a bit after process
termination, especially when another process is running in the background. As
this happens on both Windows and Linux, I suspect mymake does something strange
when interacting with the OS, but I can not figure out why.

This is a good compromise however, as we still make sure to not forget any
output before the output manager is terminated at least.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>