Fixed an issue which caused the include cache to not be saved.
On both Linux and Windows, this bug caused the include cache to be saved when
the executed process was done executing. This sometimes meant that the include
cache could be corrupted when mymake and the executed process were
killed (e.g. by Ctrl+C), since the executed process could terminate first, and
mymake could start writing include files, before getting killed in the process
of doing so.

Now, mymake actively writes the include cache before executing the finished
process instead of relying on destructors to do this.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>