Fixed the long-standing bug that caused process creation to fail on Windows.
This was since initialization of static variables was not atomic on VS2008, as I
previously presumed was the case. Now a separate lock is used to fix the issue.

The handling of failures in process creating has also been improved, so that
mymake does not end up in an infinite loop anymore.