Fixed a long-standing bug that causes mymake to crash on build failure.
For processes that produce much output and then exit, it could be the cause
that mymake would be unable to keep up with messages (e.g. due to waiting
for the terminal). If this happened during a build failure, the thread local
variable that held the output information could be deallocated, causing
both any prefixes to disappear and mymake to crash entirely. This commit
fixes both issues.